xml_lexer.h, xml_lex_0top.lex: Use 'xmlBaseFlexLexer' name and deal with multiple definition stuff.
The base class cannot be called 'yyFlexLexer' because that collides
with the base from lexer.lex.  I do not know why that was not a
problem before.

Once we do that, the multiple include guard in FlexLexer.h is
inactive, so preventing multiple definitions of that class requires
some more macro hacks.

Ideally I should just fix flex to not create this mess in the first
place.