Makefile.in, configure.pl: Minimally adapt to smbase removal of sm_config.pm.
This temporarily adds sm_config.pm and depend.pl here, but I intend
to subsequently remove them.

There are also changes related to using .exe extensions on all of the
executables.

Also I found that the tests do not pass when using -O2; that has to be
fixed.  Even without that, k0010.cc now fails, and k0010a.cc now seems
to pass but I suspect that is a bug too.