cc_print.cc: Fix the numerous indentation problems.
The old code was weirdly complicated so I rewrote it.

Also there is the unfortunate behavior that indentation and
newline buffering only happens when processing a string of
characters, not individual characters.  I just changed the
relevant characters to strings to fix it, but there is more
to be done to really fix all this.