cc_ast_aux.cc: Declarator::printExtras: print 'var' and 'type' always.
The point of 'debugPrint' (which calls printExtras) is to make the
information in the structure visible, so all fields should be
explicitly present.  In particular, the absence of 'type' was
something of an impediment during recent debugging.

Unfortunately my printouts are still inconsistent about using '='
versus ':'.  I think I should make them uniformly use '=' since
that is what astgen does, so this is one small step in that
direction.