cc_tcheck.ast: Declarator: Fix old bug: 'type' was not cloned.
The result was a Declarator whose 'var' was not NULL but whose
'type' was, which is nonsensical.

No test was failing because it is only seen if (1) there is a cast
to a structure type, and (2) the integrity check is run after
elaboration.  I'm in the process of adding commits that will
rectify these shortcomings.