ast_build: makeASTTypeId: Construct type/decl syntax instead of using TS_name.
This is part of an overall effort to remove TS_name and PQ_variable by
constructing legitimate AST to describe things rather than just pointers
to semantic structures.  This is important for source to source
translation because I need to turn these things into syntax at some
point, and it's better to do it earlier in order to avoid extra cases
later.

Add test-astbuild module and run it from main.cc (always).