cc_tcheck.ast: TypeSpecifier: Make 'clone' copy the tcheck annotations.
Generally, 'clone' should make as exact a copy as possible.  If
some client wants them to be NULL afterward, that client should
have the responsibility to clear them.  (In this case it does not
appear anything needs them cleared.)

I'm strongly tempted to make the astgen 'field' flag mean the
field is automatically copied during cloning, but I need to first
decide how that works when 'owner' is also present.