ast_build: Add 'makeParameter'.  cc_elaborate: Use it.
This is part of an ongoing effort to consolidate AST generation
in ast_build and eliminate TS_type.

This commit reorganizes a fair bit of the ASTTypeId computation.
It also leaves a couple loose ends, especially when to use TS_name
versus TS_elaborated, and when to use PQ_variable versus PQ_name.

makeASTTypeId now requires a DeclaratorContext.