Remove TS_type.
The replacement is ElsaASTBuild::makeTypeSpecifier.  I think it is
better to make real syntax, as a tree, rather than have a special
case node to carry a semantic type.  The latter increases the number
of cases that have to be handled, and we still are forced to express
the types syntactically when pretty printing.