variable: Rename Variable::scope to m_containingScope.
I want to do some maintenance on that variable, but its old name
is difficult to search for due to being so common and overloaded.

The new name conflicts with its use by namespaces to indicate the
scope they refer to, but in fact eliminating that particular
overloaded meaning is part of what I want to change.