variable: Split Variable::m_denotedScope off from m_containingScope.
The old design, in which m_containingScope (simply called 'scope') was
overloaded with two meanings, was pretty broken.  It worked (if it even
really did!) only because the limited ways in which namespace names can
be used meant that m_containingScope happened not to be used in its
"containment" meaning with namespaces, but the concept of a namespace
having a containing scope is entirely meaningful and sensible.

The basic strategy for splitting these apart is to find uses of
m_containingScope that are inside a test of 'isNamespace()', since that
is the condition where it meant something different, and change those to
m_denotedScope.