Add cc-type-visitor module.
Move all of the 'traverse(TypeVisitor&)' methods out of cc-scope,
cc-type, template, and variable into cc-type-visitor.cc.

I think the idea of a "visitor" for Types is a bit odd because
the meaning of Types is so closely tied to their structure, whereas
visitors are appropriate when meaning is mostly independent of
context.  Therefore I see TypeVisitor as more of an auxiliary
notion than a core concept, so prefer it adjacent rather than mixed
in with the rest.