cc-print: Switch from BoxPrint to TreePrint.
With this, I'm able to make test/pprint/longlines2.c work the way I
want.  longlines.cc also now looks significantly better.  pp-syntax.cc
has arguably a slight regression in how it prints function try blocks,
but that's fine and I could fix it with a bit more work.

It's still not clear whether I could have fixed this with BoxPrint,
but I think the general design in TreePrint is a bit better, and anyway
it all seems to work now, so I'll proceed with this.