cc-print: Fully print ISCs with --print-isc, suppress otherwise.
Add a parallel set of ISC-enabled pprint tests.  The regular pprint
tests now do not have any ISC stuff in them.  This separates pprint
for the purpose of evaluating layout from pprint for the purpose of
testing ISCs.

Some of the ISC output is wrong, mostly because it is incomplete.
This commit simply captures the current behavior so I can start
working on it.

Add test/pprint/stringlit.c.  The lack of ISCs there is one of the
main motivators at the moment.