cc-env: Insert E_implicitStandardConversion for all kinds of ISCs.
The old code was a patchwork of partial implementation and TODOs.
The new code should work for all kinds of conversions.

The immediate benefit is getting SC_ARRAY_TO_PTR for string literals,
which I need in the instrumentor.  The ISC tests expose the new nodes
to the extent those tests provoke them.

cc-tcheck: Additionally, this change revealed a bug in
E_implicitStandardConversion::itcheck_x involving templates, which is
now fixed.  Test in/t0216.cc is one that failed without the fix.