cc-type: Add SimpleType::getST().
This seems a bit safer since I can do a bounds check.

Change a couple uses of SimpleType::fixed[] to SimpleType::getST().
However, I left the use in the CVAT macro since that is used in the
initializer of an array and it's perhaps better to just allow the
compiler to build that all at compile time.