cc-env.cc: declareEnum: Always set 'et->typedefVar' to non-NULL.
In the existing code, anonymous CompoundTypes already had non-NULL
typedefVars, so this just makes EnumType consistent.  In the
instrumentor, I would like to be able to use the typedefVar as a
key to identify NATs, so now I can.

cc-type.h: Document that NamedAtomicType::typedefVar must not be NULL.

integrity.cc: Enforce that requirement.