cc.ast: Introduce AsmDefinition and change S_asm and TF_asm to use it.
This is a first step toward adding an AST node that holds all of
the information that the GNU 'asm' definition has.

cc.gr, gnu.gr: Construct the AsmDefinition node.

For the moment the latter still makes AD_string but my plan is to
change that next.

cc-tcheck, cc-print: Extend accordingly.

Add test/pprint/asm.c and gnu-asm.c.  The latter shows the current
(and historical) behavior of dropping the GNU info, which is what I
intend to correct next.