Allow 'complex == int'.  In C, yield 'int' for boolean operators.
This started as a simple project to allow comparison of _Complex
values to integers, which is permitted by the standard.  That is
done in gnu.cc by expanding 'prec' to allow -1 to mean integer.

But when testing this I ran into the ancient problem that Elsa
says all boolean operators yield 'bool', which is wrong in C.  So
I made changes in cc-env.cc and cc-tcheck.cc to fix that as well.

The two sets of changes are combined because testing for the former
relies on the latter.

Several C pprint tests have updated expected output due to
SC_BOOL_CONV yielding 'int' instead of 'bool' now.