cc-scope: Add trace message for each Variable registered.
Also do not print serial number or address for scope descriptions.

cc-type: CompoundType::keywordAndName: Actually print the keyword,
and print the name fully qualified with template arguments.

template: Optionally skip the invariant checks in getKind().  This
is needed while the TemplateInfo is being built for an explicit
specialization (in/t0054.cc).

test-astbuild.cc: Make the typedefVar earlier since the new
registration message needs to see it when the members are added.

variable.h: Clarify 'isGlobal()' semantics for templates.