cc-flags: Remove DF_GLOBAL.  Variable::isGlobal() is its replacement.
The now-removed assertion in Variable::isGlobal() confirmed their
equivalence on all tests.

cc-scope: There was code that tried to set DF_GLOBAL on template
instantiations, but it didn't work.  From what I could tell, it
set DF_GLOBAL on template *parameters* (like the "T" in
"template <class T>"), but that's it.  And even if it did work, I
think it is wrong to consider instantiations as "global" since
they are not found (directly) by name lookup.

variable: Remove isGlobal_flag_or_scope(), as it was equivalent to
inGlobalOrNamespaceScope().