variable: Change all hasFlag(DF_MEMBER) to isMember.  Assert equivalent to scope query.
This is a first step toward removing DF_MEMBER altogether: show
that every use of it is equivalent to querying the scopes.

This required fixing two failures to set DF_MEMBER, one in
Enumerator::tcheck and one in ND_usingDecl::tcheck.

However, I can't quite get rid of it because there is a test in
Env::createDeclaration that isn't on a Variable.