baselexer.cc: Fix possibly undefined behavior with BaseLexer ctor.
With 'gcc -fsanitize=undefined', there is a complaint about invoking a
method of an object without a valid vtable pointer.  This is due to
calling 'openFile' before the BaseLexer vtable is set.  'openFile' is
not virtual, so it works fine in practice, and after a quick scan I
could not find where in the C++ standard is a rule prohibiting, but
probably the report is accurate and this is UB.

Therefore, I've adjusted the way BaseLexer is constructed so the
'openFile' call is made only after the constructor body begins
executing.  This seems to satisfy '-fsanitize=undefined' at least.

Reported by user mingodad.